+++ naturtag/utils/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa: F401
+from naturtag.utils.i18n import read_locales
 from naturtag.utils.image_glob import get_valid_image_paths
 from naturtag.utils.thumbnails import generate_thumbnail